Adapt library and examples for 64-bit integer option (USE_INT64=ON); adapt library (but not yet examples) for 32-bit float option (USE_FP64=OFF); make gcc happy on library compilation  with -Wall -Wextra -pedantic; replace malloc/realloc/free with new/delete in container.hpp

+++ Examples/DrawGrid/color_bar.cpp
@@ -285,7 +285,7 @@ namespace INMOST
 		double px2, py2;
 		int height = glutGet(GLUT_WINDOW_HEIGHT);
 		
-		for (int i = 0; i < ticks.size() - 1; ++i)
+		for (size_t i = 0; i < ticks.size() - 1; ++i)
 		{
 			file << "<g>" << std::endl;
 			colors[i].set_color();
