+++ Source/Misc/utils.cpp
@@ -30,7 +30,7 @@ namespace INMOST {
 
     std::string string_to_lower(const std::string &str) {
         std::string lower = std::string(str);
-        for (int    i     = 0; i < lower.length(); i++)
+        for (size_t    i     = 0; i < lower.length(); i++)
             lower[i] = (char) tolower(lower[i]);
         return lower;
     }
